# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for AI-Dia application",
        "title": "AI-Dia API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8005",
    "basePath": "/api/v1",
    "paths": {
...
        "/client/{client_id}/settings/ai-prompts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all 5 AI prompt sections (product, delivery, operational, about-store, faq) for a tenant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI Prompts"
                ],
                "summary": "Get All AI Prompt Sections",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/setting.AIPromptsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/client/{client_id}/settings/ai-prompts/{section}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get AI prompt for a specific section. Valid sections: product, delivery, operational, about-store, faq",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI Prompts"
                ],
                "summary": "Get AI Prompt by Section",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "product",
                            "delivery",
                            "operational",
                            "about-store",
                            "faq"
                        ],
                        "type": "string",
                        "description": "Section name",
                        "name": "section",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/setting.AIPromptSectionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update AI prompt for a specific section. Valid sections: product, delivery, operational, about-store, faq",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI Prompts"
                ],
                "summary": "Update AI Prompt by Section",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "product",
                            "delivery",
                            "operational",
                            "about-store",
                            "faq"
                        ],
                        "type": "string",
                        "description": "Section name",
                        "name": "section",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Prompt content",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/setting.UpdateAIPromptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/client/{client_id}/settings/kds": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns whether Kitchen Display System is enabled for the given client. Accessible by SuperAdmin and the Client itself.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Get KDS status for a client",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "SuperAdmin sets whether KDS is enabled for the given client. Body: {\"enabled\": true/false}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Enable or disable KDS for a client",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "KDS toggle",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/client/{client_id}/telegram/bot-token": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Telegram bot token and auto-register webhook",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Update Telegram Bot Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Bot Token Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/telegram.UpdateBotTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/settings/integration": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get integration settings data. SuperAdmin receives all data; Client only receives Telegram data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Get Integration Settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/setting.GroupResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/settings/notification": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get notification settings data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Get Notification Settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/setting.GroupResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "helpers.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "setting.AIPromptSectionResponse": {
            "type": "object",
            "properties": {
                "prompt": {
                    "description": "The prompt text configured for this section",
                    "type": "string",
                    "example": "We are a restaurant open Sunday to Friday."
                },
                "section": {
                    "description": "Section name (product | delivery | operational | about-store | faq)",
                    "type": "string",
                    "example": "about-store"
                }
            }
        },
        "setting.AIPromptsResponse": {
            "type": "object",
            "properties": {
                "about_store": {
                    "type": "string",
                    "example": "We are a restaurant located in Jakarta."
                },
                "delivery": {
                    "type": "string",
                    "example": "We deliver to the following zones."
                },
                "faq": {
                    "type": "string",
                    "example": "1. Payment via Stripe. 2. Order expires in 15 minutes."
                },
                "operational": {
                    "type": "string",
                    "example": "Open Sunday - Friday, 07:00 - 17:00."
                },
                "product": {
                    "type": "string",
                    "example": "Show product name, price, photo, and description."
                },
                "store_operational": {
                    "type": "string",
                    "example": "{\"monday\":{\"start\":\"09:00\",\"end\":\"21:00\",\"closed\":false}}"
                }
            }
        },
...
        "setting.UpdateAIPromptRequest": {
            "type": "object",
            "properties": {
                "prompt": {
                    "description": "The prompt text for this section. Can be empty to clear the prompt.",
                    "type": "string",
                    "example": "We are a restaurant open Sunday to Friday, serving authentic local cuisine."
                }
            }
        },
...
        "telegram.UpdateBotTokenRequest": {
            "type": "object",
            "required": [
                "bot_token"
            ],
            "properties": {
                "bot_token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: \"Bearer {token}\" — paste the token from the login response (accessToken)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}